+++ test/test_database.py
@@ -739,15 +739,6 @@ class TestDatabase(IntegrationTest):
             self.assertEqual(
                 getattr(db2, opt), newopts.get(opt, getattr(db1, opt)))
 
-    def test_current_op_codec_options(self):
-        class MySON(SON):
-            pass
-        opts = CodecOptions(document_class=MySON)
-        db = self.client.get_database("pymongo_test", codec_options=opts)
-        current_op = db.current_op(True)
-        self.assertTrue(current_op['inprog'])
-        self.assertIsInstance(current_op, MySON)
-
 
 class TestDatabaseAggregation(IntegrationTest):
     def setUp(self):