+++ test/test_monitoring.py
@@ -1084,24 +1084,6 @@ class TestCommandMonitoring(IntegrationTest):
         self.assertTrue('ok' in succeeded.reply)
 
         self.listener.results.clear()
-        self.client.pymongo_test.current_op(True)
-        started = results['started'][0]
-        succeeded = results['succeeded'][0]
-        self.assertEqual(0, len(results['failed']))
-        self.assertIsInstance(started, monitoring.CommandStartedEvent)
-        expected = SON([('currentOp', 1), ('$all', True)])
-        self.assertEqualCommand(expected, started.command)
-        self.assertEqual('admin', started.database_name)
-        self.assertEqual('currentOp', started.command_name)
-        self.assertIsInstance(started.request_id, int)
-        self.assertEqual(self.client.address, started.connection_id)
-        self.assertIsInstance(succeeded, monitoring.CommandSucceededEvent)
-        self.assertIsInstance(succeeded.duration_micros, int)
-        self.assertEqual(started.command_name, succeeded.command_name)
-        self.assertEqual(started.request_id, succeeded.request_id)
-        self.assertEqual(started.connection_id, succeeded.connection_id)
-        self.assertTrue('inprog' in succeeded.reply)
-        self.assertTrue('ok' in succeeded.reply)
 
     def test_sensitive_commands(self):
         listeners = self.client._event_listeners