+++ test/test_session.py
@@ -249,7 +249,6 @@ class TestSession(IntegrationTest):
             (db.list_collection_names, [], {}),
             (db.validate_collection, ['collection'], {}),
             (db.drop_collection, ['collection'], {}),
-            (db.current_op, [], {}),
             (db.profiling_info, [], {}),
             (db.dereference, [DBRef('collection', 1)], {}),
         ]
@@ -962,14 +961,6 @@ class TestCausalConsistency(unittest.TestCase):
         # Not a write, but explain also doesn't support readConcern.
         self._test_no_read_concern(
             lambda coll, session: coll.find({}, session=session).explain())
-
-    @client_context.require_no_standalone
-    @unittest.skipIf(client_context.serverless,
-                     "Serverless does not support currentOp")
-    def test_writes_do_not_include_read_concern_current_op(self):
-        # Not a write, but currentOp also doesn't support readConcern.
-        self._test_no_read_concern(
-            lambda coll, session: coll.database.current_op(session=session))
 
     @client_context.require_no_standalone
     @unittest.skipIf(client_context.serverless,