+++ doc/changelog.rst
@@ -28,6 +28,7 @@ Breaking Changes in 4.0
   :attr:`pymongo.mongo_client.MongoClient.is_locked`.
 - Removed :meth:`pymongo.mongo_client.MongoClient.database_names`.
 - Removed :meth:`pymongo.database.Database.collection_names`.
+- Removed :meth:`pymongo.database.Database.current_op`.
 - Removed :meth:`pymongo.database.Database.authenticate` and
   :meth:`pymongo.database.Database.logout`.
 - Removed :meth:`pymongo.database.Database.error`,